+++ utils/utils.go
@@ -17,18 +17,6 @@ const (
 	dirTimestampFormat  = "20060102_150405"
 )
 
-func deepCopy(i *rpcpb.ClusterInfo) (*rpcpb.ClusterInfo, error) {
-	b, err := json.Marshal(i)
-	if err != nil {
-		return nil, err
-	}
-	var clusterInfo rpcpb.ClusterInfo
-	if err := json.Unmarshal(b, &clusterInfo); err != nil {
-		return nil, err
-	}
-	return &clusterInfo, nil
-}
-
 func ToNodeID(stakingKey, stakingCert []byte) (ids.NodeID, error) {
 	cert, err := staking.LoadTLSCertFromBytes(stakingKey, stakingCert)
 	if err != nil {
